Fixed an issue where the connection would close immediately after connecting in "net" mode.
It was possible for the client to connect to the server's tcp punch socket, which is only used temporarily, causing the connection phase to fail.
The tcp punch function has been fixed to avoid hanging, and new checks are put in place to ensure the server is ready & listening on the correct socket.